refactor: add company check for review update and delete

+++ src/main/java/me/apella/jobs/service/ReviewServiceImpl.java
@@ -1,9 +1,11 @@
 package me.apella.jobs.service;
 
+import me.apella.jobs.model.Company;
 import me.apella.jobs.model.Review;
 import me.apella.jobs.repository.ReviewRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -11,30 +13,43 @@ import java.util.UUID;
 @Service
 public class ReviewServiceImpl implements ReviewService {
     ReviewRepository reviewRepository;
+    CompanyService companyService;
 
-    public ReviewServiceImpl(ReviewRepository reviewRepository) {
+    public ReviewServiceImpl(ReviewRepository reviewRepository, CompanyService companyService) {
         this.reviewRepository = reviewRepository;
+        this.companyService = companyService;
     }
 
     @Override
     public List<Review> getAllReviews(UUID companyId) {
-        List<Review> reviews = reviewRepository.findByCompanyId(companyId);
-        return reviews;
+        Company company = companyService.getCompanyById(companyId);
+        if (company != null) {
+            List<Review> reviews = reviewRepository.findByCompanyId(companyId);
+            return reviews;
+        }
+        return new ArrayList<>();
     }
 
     @Override
-    public void createReview(UUID companyId, Review review) {
-        reviewRepository.save(review);
+    public boolean createReview(UUID companyId, Review review) {
+        Company company = companyService.getCompanyById(companyId);
+        if (company != null) {
+            review.setCompany(company);
+            reviewRepository.save(review);
+            return true;
+        }
+        return false;
     }
 
     @Override
     public Review getReviewById(UUID companyId, UUID reviewId) {
-        return reviewRepository.findById(reviewId).orElse(null);
+        List<Review> reviews = reviewRepository.findByCompanyId(companyId);
+        return reviews.stream().filter(review -> review.getId().equals(reviewId)).findFirst().orElse(null);
     }
 
     @Override
     public boolean deleteReviewById(UUID companyId, UUID reviewId) {
-        if (reviewRepository.existsById(reviewId)) {
+        if (companyService.getCompanyById(companyId) != null && reviewRepository.existsById(reviewId)) {
             reviewRepository.deleteById(reviewId);
             return true;
         }
@@ -43,14 +58,17 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public boolean updateReviewById(UUID companyId, UUID reviewId, Review updatedReview) {
-        Optional<Review> optionalReview = reviewRepository.findById(reviewId);
-        if (optionalReview.isPresent()) {
-            Review review = optionalReview.get();
-            review.setDescription(updatedReview.getDescription());
-            review.setDislikes(updatedReview.getDislikes());
-            review.setLikes(updatedReview.getLikes());
-            reviewRepository.save(review);
-            return true;
+        if (companyService.getCompanyById(companyId) != null) {
+            Optional<Review> optionalReview = reviewRepository.findById(reviewId);
+            if (optionalReview.isPresent()) {
+                Review review = optionalReview.get();
+                review.setDescription(updatedReview.getDescription());
+                review.setDislikes(updatedReview.getDislikes());
+                review.setLikes(updatedReview.getLikes());
+                reviewRepository.save(review);
+                return true;
+            }
+            return false;
         }
         return false;
     }
